app/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
joblib
pandas
scikit-learn